fix: wrongly escaped qoutation marks
a wierd thing happens when gd-rust translates the docstring to documentation. Every type of escapeable character is escaped. but that means that there is no unescaping them. This is currently resolved by using fake qoutes, but i think it conveys the example better than either to leave them out or to have all the higligthing killed.