fix: ensure that the worker has started.
I am unsure exactly why, but somehow we will get a seg fault if we do
not wait for the Actor to finish starting.

This might just be a timing issue and we have not solved the correct
thing, but just decreased the occurance of the error happening due to
more time going by before running say.

I have observed this error happening when running the following code
from C#:

-> StartWorker()
-> ResetContext() # (runs destroy and imidieately after run StartWorker
again)
-> Say()